backend/requirements.txt
fastapi==0.115.5
uvicorn==0.32.0
lxml==5.3.0
beautifulsoup4==4.12.3
pydantic==2.9.2
pytest==8.3.3
python-dotenv==1.0.1
httpx==0.27.2
html-to-markdown==2.22.2
